backend/requirements.txt
Django
django-cors-headers
openai-agents
openai
pydantic
python-dotenv
python-docx
pypdf
openpyxl

